Workaround gpu issue
For some reason starting webdriver with a user-data-dir in tmp causes
some issues when also loading some extensions - pages fail to load even
chrome pages. There are some issues reported upstream that suggest
disable the use of the gpu, the real reason is unclear, so this a hack
to work around it.