Keep running in until the driver finishes
When -S is NOT being used the bran process would exit, leaving the
profile path in place. This path was never removed,

Keep bran running in the background until the driver session finishes.
Ideally this would be done by calling wait() on the Child process api,
however the child is a private member of the driver struct (from
webdriver_client) so this method cannot be used.

Instead we poll the session url every 15s until an error takes place.