Fix bug in liveliness check loop
In e93e60e4be3a1be5a707a75c6d01cbf6a167b609 a loop was added to wait for
the driver to finish, however this was incorrect, since it failed from
non fatal error it would cause the browser to be killed. To fix this, be
more restrictive only consider exiting on Io errors and wait for at
least three errors to occur.