Fix liveliness check
Previously (4bb1eaed69b76eca5b0793b6447b8b00b29d09e2) the liveliness
check was changed because it was failling for non fatal errors.

It turns out the previous change means bran never detects the session is
finished, because the driver keeps running even if the browser shuts
down.

The previous fix happened because even if the browser is running we can
get an error e.g. the following chrome bug

    https://bugs.chromium.org/p/chromedriver/issues/detail?id=346

So instead we now use get_window_handles() which should not cause errors
in any situation.